extract alg panel from model tab
