+++ src/main/java/frontend/DatasetTab.java
@@ -39,7 +39,8 @@ public class DatasetTab extends JPanel {
 
         this.parentContext = WorkflowManager.INSTANCE.getContextByKey(Keys.App);
         this.context = new DatasetContext(parentContext, Keys.DatasetConfig);
-        this.wekaInstance = new InstanceContext(parentContext, Keys.RootWekaInstnace);
+
+        this.wekaInstance = new InstanceContext(parentContext, Keys.RootWekaInstnace);  //TODO: dataset should not live in the frontend
 
         this.constraints = new GridBagConstraints();
         this.setLayout(new GridBagLayout());
